+++ src/cpus/mos6510.rs
@@ -9,15 +9,15 @@ mod disassembler;
 
 #[derive(Debug,Default)]
 pub struct Cpu6510 {
-    pc: u16,
+    pc: u64,
     memory: BinaryBuffer
 }
 
 
 impl CpuTrait for Cpu6510 {
-    fn set_program_counter(&mut self, pc:u32) -> Result<i32,i32> {
-        if pc <= u16::MAX as u32{
-            self.pc = pc as u16;
+    fn set_program_counter(&mut self, pc:u64) -> Result<i32,i32> {
+        if pc <= u16::MAX as u64{
+            self.pc = pc;
             return Ok(0);
         }
         Err(1)
@@ -28,7 +28,7 @@ impl CpuTrait for Cpu6510 {
 impl Cpu6510 {
     pub fn new(memory: BinaryBuffer) -> Self {
         Self {
-            pc: memory.get_loaded_address() as u16,
+            pc: memory.get_loaded_address(),
             memory,
         }
     }
